Adding box2d support

+++ CoordinateHelper.cpp
@@ -1,5 +1,5 @@
 #include "precomp.h"
 #include "CoordinateHelper.h"
 
-GLint CoordinateHelper::_gameHeight = 0;
-GLint CoordinateHelper::_gameWidth = 0;
+GLfloat CoordinateHelper::_gameHeight = 0;
+GLfloat CoordinateHelper::_gameWidth = 0;
